Reduce log verbosity when compiling and linking
Instead of outputting the entire command for every file that is compiled
and linked, which is often very long and repetitive, only output a count
of the files and the root directory the files will be output to. This
mimics how SBT reports compiling Java/Scala files. For example, the
output now looks like:

  [info] compiling 10 C sources to /path/to/project/target/streams/compile/ccOutputDirectory/_global/streams/compile ...
  [info] linking 1 output files to /path/to/project/target/streams/compile/ccOutputDirectory/_global/streams/compile ...

The old output can be seen by enabling the debug log level.

This also adds a new ccOutputDirectory to reduce code duplication and to
more easily include the directory when logging.